Refactor backend message into specific type
This change should make sure that messages can be sent in a single socket write, and that fewer blocks of memory need to be allocated as the message is written to. But it should also make it easier to extend the connection to send other messages besides the ones we have built in.